% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft.R
\name{dft}
\alias{dft}
\title{Discrete Fourier Transform}
\usage{
dft(x, inverse = FALSE)
}
\arguments{
\item{x}{real or complex vector}

\item{inverse}{logical, defaults to FALSE}
}
\value{
the DFT of x
}
\description{
Discrete Fourier Transform
}
