% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{generate_mvnorm}
\alias{generate_mvnorm}
\title{Generate multivariate normal random numbers}
\usage{
generate_mvnorm(n = 10, mean_vec = c(1, 1))
}
\arguments{
\item{n}{number of observations}

\item{mean_vec}{vector of means}
}
\value{
n x length(mean_vec) of observations
from a multivariate normal
}
\description{
Generate multivariate normal random numbers
}
