% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{leeR_demo}
\alias{leeR_demo}
\title{Print Hello World!}
\usage{
leeR_demo(string)
}
\arguments{
\item{string}{character containing a string to
print}
}
\value{
A string saying "Hello World"
}
\description{
Print Hello World!
}
\examples{
leeR_demo()
}

